refactor(project): file structure, handlers
