+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "cSpell.words": [
+        "polkadot",
+        "subsquid",
+        "typegen"
+    ]
+}
