+++ src/mappings/index.ts
@@ -1,13 +1,13 @@
-import handlePostBlock from "../block/handlePostBlock";
-import handleBalancesTransfer from "../balances/handleBalancesTransfer";
-import handleTokensTransfer from "../tokens/handleTokensTransfer";
-import handlePoolCreated from "../lbp/handlePoolCreated";
-import handlePoolUpdated from "../lbp/handlePoolUpdated";
+import handlePostBlock from "./block/handlePostBlock";
+import handleBalancesTransfer from "./balances/handleBalancesTransfer";
+import handleTokensTransfer from "./tokens/handleTokensTransfer";
+import handlePoolCreated from "./lbp/handlePoolCreated";
+import handlePoolUpdated from "./lbp/handlePoolUpdated";
 
 export {
     handlePostBlock,
     handleTokensTransfer,
     handleBalancesTransfer,
     handlePoolCreated,
-    handlePoolUpdated
+    handlePoolUpdated,
 };
